deploy/requirements.txt
boto3==1.21.21
pyyaml==6.0
python-dotenv==0.19.2